# Use `rust-hash` instead of `fxhash`