`fxhash` seems not to be maintained anymore `while rustc-hash` is the "official" implementation used by rustc and maintained by rust-lang.

Did a bench just in case (1,6 GHz Intel Core i5-2467M)
before:
test bench_add_directory ... bench: 287,902,510 ns/iter (+/- 137,824,185)
test bench_probe_wdl ... bench: 3,973,610 ns/iter (+/- 2,268,762)
after:
test bench_add_directory ... bench: 271,782,730 ns/iter (+/- 75,746,895)
test bench_probe_wdl ... bench: 4,044,770 ns/iter (+/- 1,577,451)

No perf change at first sight, but the interval confidence is very large